Update Graph's class inheritance.
When `netin.DiGraph` inherits from `netin.Graph` and `nx.DiGraph`, we need to fix the order.
This way, `netin.DiGraph` has the event handling methods from `netin.Graph` but the other methods from `nx.DiGraph`.
This order causes an error with NetworkX, however.

Instead, I now change `netin.(Di)Graph` to be wrappers.
This comes at the cost of manually forwarding the NetworkX function calls, but it is flexible to the choice of network implementation.
In a future version, this could be implemented with alternative graph libraries, such as `graph-tool`.